Allow unexpected response codes to be easily checked
In conditions where previously a opaque `VaultException` would be thrown due to an unexpected response code, we now throw a `VaultResponseException`.
This exception extends `VaultException`, allowing previous error handling code to function the same way.  However it provides the additional ability to easily get at the numeric response code without needing to parse the exception message.

This can be useful to users of the library in that it allows them to have more specific error handling capabilities.